fix(segment_utils): remove uint8 truncation in cleanup_segmentation
cleanup_segmentation was casting label() output to uint8, silently
truncating labels >255 to 0. Now returns the native int32/int64 dtype
from label(). Callers that need a specific dtype already cast explicitly.

+++ cubic/segmentation/segment_utils.py
@@ -146,7 +146,7 @@ def cleanup_segmentation(
             )
             label_img[filled_mask] = label_id
 
-    return label(label_img).astype(np.uint8)
+    return label(label_img).astype(np.uint16)
 
 
 def find_objects(label_image, max_label=None):
